requirements.txt
Flask~=3.0.2
requests~=2.31.0
stripe~=8.2.0
Werkzeug~=3.0.1
WTForms~=3.1.2
Flask-WTF~=1.2.1
Flask-Login~=0.6.3
argon2-cffi~=23.1.0
